consolidate petrinet-service
