+++ packages/client/hmi-client/src/components/model/petrinet/tera-model-equation.vue
@@ -29,10 +29,9 @@ import { ref, watch } from 'vue';
 import TeraMathEditor from '@/components/mathml/tera-math-editor.vue';
 import TeraEquationContainer from '@/components/model/petrinet/tera-equation-container.vue';
 import { Model } from '@/types/Types';
-import { convertAMRToACSet } from '@/model-representation/petrinet/petrinet-service';
+import { convertAMRToACSet, petriToLatex } from '@/model-representation/petrinet/petrinet-service';
 import { equationsToAMR } from '@/services/knowledge';
 import { cleanLatexEquations } from '@/utils/math';
-import { petriToLatex } from '@/petrinet/petrinet-service';
 import { isEmpty } from 'lodash';
 import { useToastService } from '@/services/toast';
